// frontend/src/app/model/property.ts
import { DocumentDetails } from "./document-details";

export class Property {
    id: number;
    code: string;
    propertyName: string;
    contactNo: string;
    alias: string;
    address1: string;
    address2: string;
    address3: string;
    country: string;
    state: string;
    city: string;
    pincode: string;
    email: string;
    website: string;
    phone: string;
    gstNo: string;
    latitude: string;
    longitude: string;
    panNo: string;
    msmeNo: string;
    propertyLogo: string;
    propertyLogoId?: number;
    
}